+++ docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Breaking Changes
+* `ExecuteResponse` no longer returns public URLs for services in the module
 
 # 0.4.21
 ### Fixes